fixup: move TestNoGoroutineLeakWithMultipleProviders
